did some cool stuff

+++ src/game/obj_behaviors.h
@@ -163,20 +163,20 @@ s8 obj_check_if_facing_toward_angle(u32 base, u32 goal, s16 range);
 s8 obj_return_home_if_safe(struct Object *obj, f32 homeX, f32 y, f32 homeZ, s32 dist);
 void obj_spawn_yellow_coins(struct Object *obj, s8 nCoins);
 void set_object_visibility(struct Object *obj, s32 dist);
-s16 object_step(void);
-s8 is_point_within_radius_of_mario(f32 x, f32 y, f32 z, s32 dist);
+s32 object_step(void);
+s32 is_point_within_radius_of_mario(f32 x, f32 y, f32 z, s32 dist);
 void spawn_orange_number(s8 behParam, s16 relX, s16 relY, s16 relZ);
-s8 obj_flicker_and_disappear(struct Object *obj, s16 lifeSpan);
+s32 obj_flicker_and_disappear(struct Object *obj, s16 lifeSpan);
 void obj_move_xyz_using_fvel_and_yaw(struct Object *obj);
 f32 absf_2(f32 f);
-s8 is_point_close_to_object(struct Object *obj, f32 x, f32 y, f32 z, s32 dist);
+s32 is_point_close_to_object(struct Object *obj, f32 x, f32 y, f32 z, s32 dist);
 void obj_return_and_displace_home(struct Object *obj, f32 homeX, UNUSED f32 homeY, f32 homeZ, s32 baseDisp);
-s16 object_step_without_floor_orient(void);
-s16 trigger_obj_dialog_when_facing(s32 *inDialog, s16 dialogID, f32 dist, s32 actionArg);
+s32 object_step_without_floor_orient(void);
+s32 trigger_obj_dialog_when_facing(s32 *inDialog, s16 dialogID, f32 dist, s32 actionArg);
 void obj_orient_graph(struct Object *obj, f32 normalX, f32 normalY, f32 normalZ);
-s8 obj_find_wall_displacement(Vec3f dist, f32 x, f32 y, f32 z, f32 radius);
+s32 obj_find_wall_displacement(Vec3f dist, f32 x, f32 y, f32 z, f32 radius);
 void adjust_rolling_face_pitch(f32 f12);
 void set_rolling_sphere_hitbox(void);
-s8 current_mario_room_check(s16 room);
+s32 current_mario_room_check(s16 room);
 
 #endif // OBJ_BEHAVIORS_H
