+++ src/game/save_file.h
@@ -149,7 +149,7 @@ void save_file_set_cannon_unlocked(void);
 void save_file_set_cap_pos(s16 x, s16 y, s16 z);
 s32 save_file_get_cap_pos(Vec3s capPos);
 void save_file_set_sound_mode(u16 mode);
-u16 save_file_get_sound_mode(void);
+u32 save_file_get_sound_mode(void);
 void save_file_move_cap_to_default_location(void);
 void save_file_get_config(void);
 void save_file_set_config(void);
